Updated Variable and Title/Node Name constraints
Previously this said they could be "any combination of letters and numbers" but in fact they can contain letters, numbers or underscores but must begin with a letter. This has caused confusion in the past, such as with [this issue that points out node titles cannot begin with a number](https://github.com/YarnSpinnerTool/YarnSpinner/issues/280).